% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{addClusteredPoints}
\alias{addClusteredPoints}
\title{Add clustered points to a map}
\usage{
addClusteredPoints(
  map,
  coordinates,
  minDistance = 20,
  scatterDistance = 10,
  scatterRadius = 10,
  stopClusterZoom = 0.95,
  bullet = amCircle("black"),
  cluster = amCluster("orange")
)
}
\arguments{
\item{map}{an \code{amMapChart} widget}

\item{coordinates}{this can be a matrix with two columns, or a dataframe
having two columns \code{longitude} and \code{latitude} and optionally
a column \code{name} for the tooltips, or the path to a \strong{geojson}
file}

\item{minDistance}{number of pixels; bullets closer than this distance
between each other will be grouped}

\item{scatterDistance, scatterRadius, stopClusterZoom}{see
\href{https://www.amcharts.com/docs/v5/charts/map-chart/clustered-point-series/#Scatter_settings}{Scatter settings}}

\item{bullet}{list of settings for the bullets created with
\code{\link{amCircle}}, \code{\link{amTriangle}} or \code{\link{amRectangle}}}

\item{cluster}{list of settings for the clusters created with
\code{\link{amCluster}}}
}
\value{
An \code{amMapChart} widget.
}
\description{
Add some clustered points to an \code{amMapChart} (points close
  to each other are grouped).
}
\examples{
library(amMapCharts5)
continents <-
  system.file("geojson", "continentsLow.json", package = "amMapCharts5")
amMapChart() |>
  addPolygons(continents, color = "violet", strokeColor = "black") |>
  addClusteredPoints(
    cities,
    bullet = amTriangle("red", strokeColor = "black"),
    cluster = amCluster("orange")
  )
}
