% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{addLineWithPlane}
\alias{addLineWithPlane}
\title{Add a line with a plane to a map}
\usage{
addLineWithPlane(
  map,
  coordinates,
  planePosition = 0.5,
  lineType = "curved",
  color = "red",
  opacity = 1,
  width = 3
)
}
\arguments{
\item{map}{an \code{amMapChart} widget}

\item{coordinates}{coordinates for a single line, a numeric matrix with
two columns or a dataframe having two columns \code{longitude} and
\code{latitude}}

\item{planePosition}{relative position of the plane on the line
(0: beginning, 1: end)}

\item{lineType, color, opacity, width}{see \code{\link{addLines}}}
}
\value{
An \code{amMapChart} widget.
}
\description{
Add a line with a plane on it to an \code{amMapChart}.
}
\examples{
library(amMapCharts5)
continents <-
  system.file("geojson", "continentsLow.json", package = "amMapCharts5")
line <- rbind(
  c(-73.778137, 40.641312),
  c(-0.454296, 51.470020),
  c(116.597504, 40.072498)
)
amMapChart(projection = "naturalEarth1") \%>\%
  addPolygons(continents, color = "red", strokeColor = "black") \%>\%
  addLineWithPlane(line, planePosition = 0.2, color = "lime", width = 2)
}
