% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{addPolygons}
\alias{addPolygons}
\title{Title}
\usage{
addPolygons(
  map,
  coordinates,
  color = NULL,
  opacity = NULL,
  strokeColor = NULL,
  strokeWidth = NULL
)
}
\arguments{
\item{map}{an \code{amMapChart} widget}

\item{coordinates}{for a single polygon, this can be a numeric matrix with
two columns or a dataframe having two columns \code{longitude} and
\code{latitude}; for multiple polygons, a list of such matrices or
dataframes; or the path to a \strong{geojson} file}

\item{color}{fill color}

\item{opacity}{opacity, a number between 0 and 1}

\item{strokeColor}{stroke color}

\item{strokeWidth}{stroke width}
}
\value{
An \code{amMapChart} widget.
}
\description{
Title
}
\examples{
library(amMapCharts5)
continents <-
  system.file("geojson", "continentsLow.json", package = "amMapCharts5")
amMapChart() \%>\%
  addPolygons(continents, color = "orange", strokeColor = "black")
}
