% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullets.R
\name{amCluster}
\alias{amCluster}
\title{Cluster settings}
\usage{
amCluster(color = NULL, radius = 6, labelColor = "black", fontSize = 9)
}
\arguments{
\item{color}{color}

\item{radius}{radius}

\item{labelColor}{label color}

\item{fontSize}{label font size}
}
\value{
A list of settings for the clusters.
}
\description{
Settings of clusters for usage in
  \code{\link{addClusteredPoints}}.
}
