% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amMapChart.R
\name{amMapChart-shiny}
\alias{amMapChart-shiny}
\alias{amMapChartOutput}
\alias{renderAmMapChart}
\title{Shiny bindings for 'amMapChart'}
\usage{
amMapChartOutput(outputId, width = "100\%", height = "400px")

renderAmMapChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{a valid CSS dimension (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates an \code{\link{amMapChart}}}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{logical, whether \code{expr} is a quoted expression
(with \code{quote()}); this is useful if you want to save an expression
in a variable}
}
\value{
\code{amMapChartOutput} returns an output element that can be
  included in a Shiny UI definition, and \code{renderAmMapChart} returns a
  \code{shiny.render.function} object that can be included in a Shiny server
  definition.
}
\description{
Output and render functions for using \code{amMapChart} within
  Shiny applications and interactive Rmd documents.
}
