% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amMapChart.R
\name{amMapChart}
\alias{amMapChart}
\title{Chart widget}
\usage{
amMapChart(
  projection = "Mercator",
  grid = list(step = 10, color = "black", opacity = 0.1),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{projection}{the projection xxx}

\item{grid}{list of settings for the grid; set to \code{NULL} for no grid}

\item{width, height}{dimensions}

\item{elementId}{a HTML id}
}
\value{
An \code{amMapChart} widget.
}
\description{
Initiate an \code{amMapChart} widget.
}
\examples{
library(amMapCharts5)
world <-
  system.file("geojson", "worldLow.json", package = "amMapCharts5")
amMapChart(projection = "orthographic") \%>\%
  addPolygons(world, color = "red", strokeColor = "black")
}
