% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{amMapCharts5-imports}
\alias{amMapCharts5-imports}
\alias{JS}
\alias{saveWidget}
\alias{\%>\%}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
  Follow the links to their documentation:
  \code{\link[htmlwidgets:JS]{JS}},
  \code{\link[htmlwidgets:saveWidget]{saveWidget}},
  \code{\link[magrittr:\%>\%]{\%>\%}}.
}
