% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullets.R
\name{amMapCharts5-shapes}
\alias{amMapCharts5-shapes}
\alias{amTriangle}
\alias{amCircle}
\alias{amRectangle}
\title{Bullets}
\usage{
amTriangle(
  color = NULL,
  opacity = 1,
  width = 10,
  height = 10,
  strokeColor = NULL,
  strokeOpacity = 1,
  strokeWidth = 2,
  rotation = 0
)

amCircle(
  color = NULL,
  opacity = 1,
  radius = 4,
  strokeColor = NULL,
  strokeOpacity = 1,
  strokeWidth = 2
)

amRectangle(
  color = NULL,
  opacity = 1,
  width = 10,
  height = 10,
  strokeColor = NULL,
  strokeOpacity = 1,
  strokeWidth = 2,
  rotation = 0,
  cornerRadius = 3
)
}
\arguments{
\item{color}{bullet color}

\item{opacity}{bullet opacity, a number between 0 and 1}

\item{width}{bullet width}

\item{height}{bullet height}

\item{strokeColor}{stroke color of the bullet}

\item{strokeOpacity}{stroke opacity of the bullet, a number between 0 and 1}

\item{strokeWidth}{stroke width of the bullet}

\item{rotation}{rotation angle}

\item{radius}{circle radius}

\item{cornerRadius}{radius of the rectangle corners}
}
\value{
A list of settings for the bullets.
}
\description{
Create a list of settings for bullets,
  their shape and their style.
}
\note{
A color can be given by the name of a R color, the name of a CSS
  color, e.g. \code{"indigo"} or \code{"fuchsia"}, an HEX code like
  \code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
  like \code{"hsl(360,11,255)"}.
}
