% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cities.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities}
\format{
A dataframe with 165 rows and 3 columns.
}
\usage{
cities
}
\description{
Cities data with name, latitude and longitude.
}
\keyword{datasets}
