% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{addLines}
\alias{addLines}
\title{Add lines to a map}
\usage{
addLines(
  map,
  coordinates,
  lineType = "curved",
  color = NULL,
  opacity = NULL,
  width = NULL
)
}
\arguments{
\item{map}{an \code{amMapChart} widget}

\item{coordinates}{for a single line, this can be a numeric matrix with
two columns or a dataframe having two columns \code{longitude} and
\code{latitude}; for multiple lines, a list of such matrices or
dataframes; or the path to a \strong{geojson} file}

\item{lineType}{type of the line, \code{"curved"} or \code{"straight"}}

\item{color}{line color}

\item{opacity}{line opacity, a number between 0 and 1}

\item{width}{line width}
}
\value{
An \code{amMapChart} widget.
}
\description{
Add some lines to an \code{amMapChart}.
}
\note{
A color can be given by the name of a R color, the name of a CSS
  color, e.g. \code{"lime"} or \code{"indigo"}, an HEX code like
  \code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
  like \code{"hsl(360,11,255)"}.
}
\examples{
library(amMapCharts5)
continents <-
  system.file("geojson", "continentsLow.json", package = "amMapCharts5")
line <- rbind(
  c(-73.778137, 40.641312),
  c(-0.454296, 51.470020),
  c(116.597504, 40.072498)
)
amMapChart() \%>\%
  addPolygons(continents, color = "orange", strokeColor = "black") \%>\%
  addLines(line, color = "purple", width = 3)
}
