% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{addPolygons}
\alias{addPolygons}
\title{Add polygons to a map}
\usage{
addPolygons(
  map,
  coordinates,
  tooltipKey = NULL,
  color = NULL,
  opacity = NULL,
  strokeColor = NULL,
  strokeWidth = NULL
)
}
\arguments{
\item{map}{an \code{amMapChart} widget}

\item{coordinates}{for a single polygon, this can be a numeric matrix with
two columns or a dataframe having two columns \code{longitude} and
\code{latitude}; for multiple polygons, a list of such matrices or
dataframes; or the path to a \strong{geojson} file}

\item{tooltipKey}{if \code{coordinates} is given by a \strong{geojson} file,
the name of a property whose value will appear in tooltips}

\item{color}{fill color}

\item{opacity}{opacity, a number between 0 and 1}

\item{strokeColor}{stroke color}

\item{strokeWidth}{stroke width}
}
\value{
An \code{amMapChart} widget.
}
\description{
Add some polygons to an \code{amMapChart}.
}
\note{
A color can be given by the name of a R color, the name of a CSS
  color, e.g. \code{"crimson"} or \code{"fuchsia"}, an HEX code like
  \code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
  like \code{"hsl(360,11,255)"}.
}
\examples{
library(amMapCharts5)
world <-
  system.file("geojson", "worldLow.json", package = "amMapCharts5")
amMapChart() \%>\%
  addPolygons(
    world, tooltipKey = "name", color = "red", strokeColor = "black"
  )
}
