% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amMapChart.R
\name{amMapChart}
\alias{amMapChart}
\title{Chart widget}
\usage{
amMapChart(
  projection = "Mercator",
  panX = "translateX",
  panY = "translateY",
  grid = list(step = 10, color = "black", opacity = 0.2),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{projection}{the projection, one choice among \code{"equalEarth"},
\code{"equirectangular"}, \code{"Mercator"}, \code{"naturalEarth1"} or
\code{"orthographic"}}

\item{panX, panY}{see \href{https://www.amcharts.com/docs/v5/charts/map-chart/map-pan-zoom/#Panning}{Panning}}

\item{grid}{list of settings for the grid; set to \code{NULL} for no grid}

\item{width, height}{dimensions}

\item{elementId}{a HTML id (usually useless)}
}
\value{
An \code{amMapChart} widget.
}
\description{
Initiates an \code{amMapChart} widget.
}
\examples{
library(amMapCharts5)
world <-
  system.file("geojson", "worldLow.json", package = "amMapCharts5")
amMapChart("orthographic", panX = "rotateX", panY = "rotateY") \%>\%
  addPolygons(world, color = "red", strokeColor = "black")
}
