% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amMapChart.R
\name{amMapChart}
\alias{amMapChart}
\title{<Add Title>}
\usage{
amMapChart(
  projection = "Mercator",
  grid = list(step = 10, color = "black", opacity = 0.1),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
